Make a directory for score functions.
Previously we were trying to treat these like flags, but there wasn't
really a good way to specify a path to a weights file from a flag file.
Now I've added support for custom score functions directly to PIP, so it
makes sense to keep the score functions in their own place.