Renumber and correct the resfiles.
I was worried about renumbering the resfiles to match the new input
structure, because they have so many residues that I could so easily get
wrong.  So I wrote a script to help me do this.  It needs a residue map,
and I made one by opening the new and old structures in pymol and
writing interactive python/pymol code.  I had to be very careful about
this step, and it took me a few tries to get it right.

While doing this, I realized that these resfiles I'd previously
committed were not the ones we used the first time through.  So I found
the right resfiles (for designing both 11 and 15 residues), copied them
in, renumbered that, and that's the commit.